refactor tests for project awesome api
